Moved game creation form to its own page.
* Game creation form is now at URL /create
* Added Create Game to nav bar.
* Removed tabbed interface from home page now that it contains only one tab's
  worth of content.

Resolves #175
