+++ imports/ui/create.html
@@ -1,0 +1,45 @@
+<template name="create_game">
+  <h1>Create Game</h1>
+
+  <div class="form-group">
+    <label for="name-input">Name</label>
+    <input class="form-control" id="name-input" type="text">
+  </div>
+
+  <div id="select-opp-div">
+    <label>Opponents</label>
+    {{#each aiPlayers}}
+      <div class="radio">
+        <label><input type="radio" name="opp-radio" value="{{name}}">{{full_name}}</label>
+      </div>
+    {{/each}}
+    <div class="radio">
+      <label><input type="radio" name="opp-radio" value="waiting">Post to Waiting Room</label>
+    </div>
+    <div class="radio">
+      <label><input type="radio" name="opp-radio" value="friend">Friend</label>
+    </div>
+  </div>
+  <br />
+  {{#if isOpponentAi}}
+    {{#with selectedAi}}
+      <p><h3>{{full_name}}</h3></p>
+      <p><span><strong>Difficulty: </strong>{{difficulty_name}}</span></p>
+      <p><span><strong>Description: </strong>{{description}}</span></p>
+    {{/with}}
+  {{/if}}
+  {{#if isOpponentFriend}}
+    <div class="form-group">
+      <label for="friend-input">Friend Account</label>
+      <input class="form-control" id="friend-input" type="text">
+    </div>
+  {{/if}}
+
+  {{#if currentUser}}
+    <button id="createGame" type="submit" class="btn btn-primary btn-md">
+      Create Game
+    </button>
+  {{else}}
+    Log in to create a game.
+  {{/if}}
+</template>
