+++ imports/ui/home.js
@@ -23,21 +23,6 @@ Template.home.helpers({
   },
 });
 
-Template.game_create_form.helpers({
-  aiPlayers() {
-    return AI.getPlayers("difficulty");
-  },
-  isOpponentAi() {
-    return Session.get('isOpponentAI');
-  },
-  isOpponentFriend() {
-    return Session.get('isOpponentFriend');
-  },
-  selectedAi() {
-    return Session.get('selectedAI');
-  }
-});
-
 Template.home_game.helpers({
   gameState(gameId) {
     const state = Games.findOne({_id: gameId}, {state: 1});
@@ -72,80 +57,9 @@ Template.home_game.helpers({
   },
 });
 
-Template.home.events({
-  'click #createGame'() {
-    const user = Meteor.user();
-    if(!user) throw new Meteor.error('not-logged-in');
-
-    let gotoGame = true;
-    var game;
-
-    if (Session.get('isOpponentAI'))
-    {
-      game = Game.create(user, null, $("#name-input").val());
-      Game.initVsAi(game, Session.get('selectedAI').name);
-    }
-    else if (Session.get('isOpponentFriend'))
-    {
-      const oppUsername = $("#friend-input").val();
-
-      const oppUser = Meteor.users.findOne( { 'username': oppUsername } );
-
-      if (oppUser === null || oppUser === undefined)
-      {
-        console.log("User doesn't exist...");
-        gotoGame = false;
-      }
-      else if (oppUser._id === user._id)
-      {
-        console.log("Can't play against yourself...");
-        gotoGame = false;
-      }
-      else
-      {
-        game = Game.create(user, null, $("#name-input").val());
-        Game.joinWaiting(game, oppUser);
-      }
-    }
-    else  // Post to waiting room
-    {
-      game = Game.create(user, null, $("#name-input").val());
-      Game.initToWaiting(game);
-    }
-
-    if (gotoGame)
-    {
-      FlowRouter.go('game', { id: game._id });
-    }
-  }
-});
-
 Template.home_game.events({
   'click .deleteGame'(event) {
     event.preventDefault();
     Games.remove(this._id);
   }
-});
-
-Template.game_create_form.events({
-  'change #select-opp-div'() {
-    const oppSelection = $("input[name='opp-radio']:checked").val();
-
-    if (oppSelection === 'friend')
-    {
-      Session.set('isOpponentAI', false);
-      Session.set('isOpponentFriend', true);
-    }
-    else if (oppSelection === 'waiting')
-    {
-      Session.set('isOpponentAI', false);
-      Session.set('isOpponentFriend', false);
-    }
-    else
-    {
-      Session.set('isOpponentAI', true);
-      Session.set('isOpponentFriend', false);
-      Session.set('selectedAI', AI.getPlayer(oppSelection));
-    }
-  },
 });
