+++ imports/ui/layout.html
@@ -15,6 +15,7 @@
         <ul class="nav navbar-nav">
           <li><a href="/">Home</a></li>
           <li><a href="/rules">Rules</a></li>
+          <li><a href="/create">Create Game</a></li>
           <li><a href="/waiting">Waiting Room</a></li>
         </ul>
         <ul class="nav navbar-nav navbar-right">
